backend/requirements.txt
# Phase 2 Backend Dependencies
fastapi>=0.110.0
uvicorn[standard]>=0.29.0
supabase>=2.3.4
python-multipart>=0.0.9
pydantic-settings>=2.2.1
pyjwt>=2.8.0
python-dotenv>=1.0.1

# Phase 3 AI Dependencies
torch>=2.0.0
torchvision>=0.15.0
transformers>=4.38.0
Pillow>=10.2.0
insightface>=0.7.3
onnxruntime>=1.17.0
sentence-transformers>=2.5.1
